fix(ai): move filterFreeModels to utils, add :free model enforcement test
- Extract filterFreeModels and OpenRouterModel type to filter-utils.ts so
  that route.ts only exports HTTP handlers (Next.js 15 prohibits arbitrary
  named exports from route files — build was failing with
  '"filterFreeModels" is not a valid Route export field')
- Remove next: { revalidate: 3600 } from fetch() options (not valid
  RequestInit; Next.js fetch caching not needed for an authed API route)
- Update route test to import filterFreeModels from ../filter-utils and
  drop the revalidate assertion from the fetch call check
- Add missing test: openrouter_free rejects non-:free model IDs (400)
  to cover the P1 review concern about paid model bypass
